ISSUE#106
Disabling buttons after user interaction
