+++ app/src/main/java/tobedevelopers/project_fury/backlog/BacklogContract.java
@@ -1,5 +1,7 @@
 package tobedevelopers.project_fury.backlog;
 
+import tobedevelopers.project_fury.backlog.implementation.Holder;
+
 /**
  * Created by Macro303 on 12/08/2016.
  */
