+++ app/src/main/java/tobedevelopers/project_fury/backlog/implementation/BacklogView.java
@@ -12,8 +12,6 @@ import tobedevelopers.project_fury.BaseNavigationView;
 import tobedevelopers.project_fury.R;
 import tobedevelopers.project_fury.ToastLog;
 import tobedevelopers.project_fury.backlog.BacklogContract;
-import tobedevelopers.project_fury.backlog.BacklogFragmentPagerAdapter;
-import tobedevelopers.project_fury.backlog.Holder;
 import tobedevelopers.project_fury.create_task.implementation.CreateTaskView;
 import tobedevelopers.project_fury.runnable_param.Runnable1Param;
