+++ app/src/main/java/tobedevelopers/project_fury/create_task/implementation/CreateTaskView.java
@@ -72,6 +72,7 @@ public class CreateTaskView extends BaseView implements CreateTaskContract.View,
 	public void onUserSelectAButton( View view ){
 		switch( view.getId() ){
 			case R.id.createTaskActivity_createTaskButton:
+				mCreateTaskButton.setEnabled( false );
 				ToastLog.makeDebug( this, "Create Task", Toast.LENGTH_SHORT );
 				presenter.userSelectCreateTask( mAssigneeSpinner.getSelectedItem().toString() );
 				break;
@@ -175,11 +176,13 @@ public class CreateTaskView extends BaseView implements CreateTaskContract.View,
 
 	@Override
 	public void noInternetAccessValidation(){
+		mCreateTaskButton.setEnabled( true );
 		ToastLog.makeWarn( this, getString( R.string.error_noInternetAccess ), Toast.LENGTH_LONG );
 	}
 
 	@Override
 	public void errorValidation(){
+		mCreateTaskButton.setEnabled( true );
 		ToastLog.makeWarn( this, getString( R.string.error_defaultError ), Toast.LENGTH_LONG );
 	}
 }
