+++ app/src/main/java/tobedevelopers/project_fury/dashboard/implementation/DashboardView.java
@@ -18,10 +18,6 @@ import tobedevelopers.project_fury.R;
 import tobedevelopers.project_fury.ToastLog;
 import tobedevelopers.project_fury.create_project.implementation.CreateProjectView;
 import tobedevelopers.project_fury.dashboard.DashboardContract;
-import tobedevelopers.project_fury.dashboard.ProjectAdapter;
-import tobedevelopers.project_fury.dashboard.ProjectHolder;
-import tobedevelopers.project_fury.dashboard.TaskAdapter;
-import tobedevelopers.project_fury.dashboard.TaskHolder;
 import tobedevelopers.project_fury.model.Model;
 import tobedevelopers.project_fury.model.Project;
 import tobedevelopers.project_fury.model.Task;
