+++ app/src/main/java/tobedevelopers/project_fury/dashboard/implementation/ProjectHolder.java
@@ -1,4 +1,4 @@
-package tobedevelopers.project_fury.backlog;
+package tobedevelopers.project_fury.dashboard.implementation;
 
 import java.util.HashMap;
 
@@ -7,15 +7,15 @@ import tobedevelopers.project_fury.model.Project;
 import tobedevelopers.project_fury.model.Task;
 
 /**
- * Created by Macro303 on 12/09/2016.
+ * Created by Macro303 on 7/09/2016.
  */
-public class Holder{
+public class ProjectHolder{
 
 	private Project[] projects;
 	private HashMap< String, Task[] > tasks;
 	private HashMap< String, Column[] > columns;
 
-	public Holder( Project[] projects ){
+	public ProjectHolder( Project[] projects ){
 		this.projects = projects;
 		tasks = new HashMap<>();
 		columns = new HashMap<>();
@@ -25,19 +25,19 @@ public class Holder{
 		this.tasks.put( projectName, tasks );
 	}
 
-	public Project[] getProjects(){
-		return projects;
-	}
-
-	public Task[] getTasks( String projectName ){
-		return tasks.get( projectName );
-	}
-
 	public void addColumns( String projectName, Column[] columns ){
 		this.columns.put( projectName, columns );
 	}
 
-	public Column[] getColumns( String projectName ){
-		return columns.get( projectName );
+	public HashMap< String, Column[] > getColumns(){
+		return columns;
+	}
+
+	public Project[] getProjects(){
+		return projects;
+	}
+
+	public HashMap< String, Task[] > getTasks(){
+		return tasks;
 	}
 }
