+++ app/src/main/java/tobedevelopers/project_fury/login/implementation/LoginPresenter.java
@@ -67,6 +67,7 @@ public class LoginPresenter implements LoginContract.Presenter{
 							break;
 						default:
 							view.setInvalidUserValidation();
+							view.enableLoginButton();
 							break;
 					}
 				}
